@core.next: code deduplication
